# monolith/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"AuthControllerType": {
				"properties": {
					"firstname": {
						"type": "string"
					},
					"lastname": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"firstname",
					"lastname",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Authlogin": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "learnwithkru",
		"version": "1.0.0",
		"description": "Learnwithkru api project",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Smoeury Songvat"
		}
	},
	"paths": {
		"/api/v1/auth/signup": {
			"post": {
				"operationId": "CreateAuth",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthControllerType"
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/verify": {
			"get": {
				"operationId": "VerifyEmail",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/auth/login": {
			"post": {
				"operationId": "LoginWithEmail",
				"responses": {
					"201": {
						"description": "CREATE"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Authlogin"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}